섹션4 38 완료

+++ site/src/commonMain/kotlin/com/kake/base/models/Post.kt
@@ -3,7 +3,7 @@ package com.kake.base.models
 expect class Post {
     val id: String
     val author: String
-    val date: Long
+    val date: Double
     val title: String
     val subtitle: String
     val thumbnail: String
